Homepage: Move App component into pages dir
