+++ apps/homepage/src/pages/App/App.scss
@@ -1,6 +1,6 @@
 // @use "../node_modules/bootstrap/scss/vendor/_rfs.scss";
 @import "./../../sass/variables.scss";
-@import "./../FontOptions/fontOptions.scss";
+@import "./../../components/FontOptions/fontOptions.scss";
 @import "~bootstrap/scss/bootstrap.scss";
 
 // App Styling
